+++ rllib/core/learner/learner.py
@@ -1143,7 +1143,8 @@ class Learner(Checkpointable):
         *,
         timesteps: Optional[Dict[str, Any]] = None,
         minibatch_size: Optional[int] = None,
-        num_iters: int = 1,
+        num_iters: int = None,
+        **kwargs,
     ):
         self._check_is_built()
         minibatch_size = minibatch_size or 32
@@ -1162,8 +1163,7 @@ class Learner(Checkpointable):
         for batch in iterator.iter_batches(
             batch_size=minibatch_size,
             _finalize_fn=_finalize_fn,
-            prefetch_batches=2,
-            local_shuffle_buffer_size=minibatch_size * 10,
+            **kwargs,
         ):
             # Update the iteration counter.
             i += 1
